requirements.txt
rapidfuzz
transformers
torch
tf-keras
fastapi
groq
python-dotenv
uvicorn
torch-geometric
cryptography
pymongo
scikit-learn